+++ src/components/Alert/interfaces.ts
@@ -1,0 +1,5 @@
+import { HTMLAttributes, ReactNode } from 'react';
+export interface IAlertProps extends HTMLAttributes<HTMLDivElement> {
+  variant?: 'danger' | 'warning' | 'info';
+  children?: ReactNode;
+}